Update and fix Homepage Hero styling & behaviors
I really didn't appreciate the flashing on the titles in the hero so
I am now binding jQuery events mouseenter and mouseleave. They handle
the parents checking for me and I'm okay with that. I also tweaked the
CSS transitions by keeping them on the element instead of the active
state. I'm doing some smarter interval checking to ensure that I'm not
running the timer more than necessary. And, originally I had forgotten
to remove the interval when you click or tap the And? button which
caused the interval to stack. But that's now fixed.